# allow creating histograms without lookup tables