@stevekuznetsov 2021-06-07
@postwait I'm still thinking through the simplest way to expose this type of space-efficient histogram to users that implicitly deserialize using JSON, but this patch stands alone to allow users to create and manipulate histograms without the lookup table manually.