@stevekuznetsov 2021-06-07
I think I've also forgotten to show my work w.r.t. space utilization. I have a JSON file with 38062 small-ish histograms on my laptop, total size on disk is 3MiB. If I simply load this using `json.Unmarshal` before this patch, I use up ~600MiB, about 95% of which is allocating the `Histogram` structures, which are made up mostly of the lookup tables:

```
$ go tool pprof -top -alloc_space /tmp/heap-profiles-735465342/heap-profile-633166655
File: pod-scaler
Type: alloc_space
Time: Jun 7, 2021 at 4:02pm (PDT)
Showing nodes accounting for 584.24MB, 98.59% of 592.61MB total
Dropped 30 nodes (cum <= 2.96MB)
      flat  flat%   sum%        cum   cum%
  235.94MB 39.81% 39.81%   235.94MB 39.81%  github.com/openhistogram/circonusllhist.New
  219.86MB 37.10% 76.91%   219.86MB 37.10%  reflect.New
   57.10MB  9.63% 86.55%    57.10MB  9.63%  github.com/openhistogram/circonusllhist.(*Histogram).insertNewBinAt
   48.02MB  8.10% 94.65%    48.02MB  8.10%  github.com/openhistogram/circonusllhist.(*Histogram).updateFast (inline)
```